refactor(test): moved jest d3 mocks to __mocks__
