+++ pages/src/pages/line.page.ts
@@ -1,3 +1,4 @@
+import { maxValueCustomMode } from "../../../tests/unit/__utils__/utils";
 import { TestPageBundleFactory } from "../types";
 import { renderPage } from "../utils";
 
@@ -11,34 +12,7 @@ export default (({ isE2ETest }) => [
 		title: "Line (linear, custom interaction mode)",
 		fileName: "line-linear-custom-interaction.html",
 		isE2ETest: isE2ETest,
-		customJS: `
-			Chart.Interaction.modes.maxValueCustomMode = function (
-				chart,
-				e,
-				options,
-				useFinalPosition,
-			) {
-				const nearestItems = Chart.Interaction.modes.nearest(
-					chart,
-					e,
-					{ axis: "x", intersect: false },
-					useFinalPosition,
-				);
-
-				let maxItem = null,
-					maxValue = -Infinity;
-				for (const item of nearestItems) {
-					// console.log(item, chart.data.datasets[item.datasetIndex])
-					const value = chart.data.datasets[item.datasetIndex].data[item.index].y;
-					if (value > maxValue) {
-						maxValue = value;
-						maxItem = item;
-					}
-				}
-
-				return maxItem ? [maxItem] : [];
-			};
-		`,
+		customJS: `Chart.Interaction.modes.maxValueCustomMode = ${maxValueCustomMode.toString()}`,
 		// make only the y-axis draggable in demo mode; E2E tests are not carried out for this page
 		demoOnlyDraggableAxis: "y",
 	}),
