+++ tests/unit/__mocks__/d3-drag.ts
@@ -1,0 +1,7 @@
+module.exports = {
+	drag: jest.fn(() => ({
+		container: jest.fn().mockReturnThis(),
+		on: jest.fn().mockReturnThis(),
+		apply: jest.fn().mockReturnThis(),
+	})),
+};
